pyproject.toml
[project]
name = "git_toxic"
version = "0.1.0"
dependencies = ["pyfswatch @ git+https://github.com/paul-nameless/pyfswatch"]

[project.optional-dependencies]
dev = ["mypy"]

[project.scripts]
git-toxic = "git_toxic:entry_point"

[tool.mypy]
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"
files = ["src"]
strict = true
warn_unreachable = true
extra_checks = true
enable_error_code = "possibly-undefined, redundant-expr, truthy-iterable, truthy-bool"

[[tool.mypy.overrides]]
module = ["fswatch.*"]
ignore_missing_imports = true

[tool.black]
target-version = ["py312"]
skip-magic-trailing-comma = true

[tool.isort]
profile = "black"
force_single_line = true
